Update Clang-Format specification
The new formatting is based on Google style, almost all configuration statements regarding C/C++ code are contained. There are certain issues with this format because clang-format is lacking some features to produce a better looking result, some comments have been added to affecting options.

Because of these issues it is currently not advised to reformat the code base according to this specification.